1/Method to utils
